% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_description}
\alias{add_description}
\alias{add_description.data.frame}
\alias{add_description.metaframe}
\title{Add a description to data in R}
\usage{
add_description(data, src)

\method{add_description}{data.frame}(data, descr)

\method{add_description}{metaframe}(data, descr)
}
\arguments{
\item{data}{a dataset to append with source information}

\item{descr}{a character vector describing the data}
}
\description{
This function will add a character description to a data object
in R. If a description already exists, it will be replaced. If a description
 does not exist, it will convert the object to a metaframe, and append the
 description

Add a description to a data.frame

Add a description to a data.frame
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: 

\item \code{metaframe}: 
}}

