% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_label}
\alias{add_label}
\alias{add_label.meta.data}
\alias{add_label.metaframe}
\title{Add a label to meta.data in R}
\usage{
add_label(object, label)

\method{add_label}{meta.data}(object, label)

\method{add_label}{metaframe}(object, label)
}
\arguments{
\item{object}{an object to append with meta.data labels}

\item{label}{a character vector or list for the labels}
}
\description{
This function will add a description to a data object
in R. If a description already exists, it will be replaced. If a description
 does not exist, it will convert the object to a metaframe, and append the
 description
}
