% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_note}
\alias{add_note}
\alias{add_note.meta.data}
\alias{add_note.meta.frame}
\title{Add annotation information to data in R}
\usage{
add_note(object, note)

\method{add_note}{meta.data}(object, note)

\method{add_note}{meta.frame}(object, note)
}
\arguments{
\item{object}{an object to append with meta.data notes}

\item{note}{a character vector or list for the notes}
}
\description{
This function will add annotations to the dataset. Annotations
should be non-data descriptions of dataset features such as observation dates,
log information, or equipment changes in collecting data.

Add a note to a meta.data

Add notes to a meta.frame
}
\section{Methods (by class)}{
\itemize{
\item \code{meta.data}: 

\item \code{meta.frame}: 
}}

