% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_revision}
\alias{add_revision}
\alias{add_revision.meta.data}
\alias{add_revision.meta.frame}
\title{Add revisions information to data in R}
\usage{
add_revision(object, revision)

\method{add_revision}{meta.data}(object, revision)

\method{add_revision}{meta.frame}(object, revision)
}
\arguments{
\item{object}{an object to append with meta.data about revisions to the data}

\item{revision}{a character vector or list of revisions}
}
\description{
This function will add revision information to the dataset. Revisions
are descriptions of changes made to individual fields in the dataset based on
information or changes outside of the analysis tool-chain

Add a revisions to meta.data

Add revisions to a meta.frame
}
\section{Methods (by class)}{
\itemize{
\item \code{meta.data}: 

\item \code{meta.frame}: 
}}

