% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_revisions}
\alias{add_revisions}
\alias{add_revisions.data.frame}
\alias{add_revisions.metaframe}
\title{Add information about revisions to data in R}
\usage{
add_revisions(data, revs)

\method{add_revisions}{data.frame}(data, revs)

\method{add_revisions}{metaframe}(data, revs)
}
\arguments{
\item{data}{a dataset to modify}

\item{revs}{a list containing descriptions of revisions}
}
\description{
This function will add information about revisions to a dataset.
Revisions are descriptions of changes made to the dataset including  changing
modifications to values of a particular row, column, or cell.

Add revisions to a data.frame

Add revisions to a metaframe
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: 

\item \code{metaframe}: 
}}

