% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_unit}
\alias{add_unit}
\alias{add_unit.meta.data}
\alias{add_unit.meta.frame}
\title{Add units information to data in R}
\usage{
add_unit(object, unit)

\method{add_unit}{meta.data}(object, unit)

\method{add_unit}{meta.frame}(object, unit)
}
\arguments{
\item{object}{an object to append with meta.data about units of measurement}

\item{note}{a character vector of list for the units}
}
\description{
This function will add units information to the dataset. Units
should be short character descriptions of the unit of measure for data elements
in the data set.

Add a unit to a meta.data

Add units to a meta.frame
}
\section{Methods (by class)}{
\itemize{
\item \code{meta.data}: 

\item \code{meta.frame}: 
}}

