% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\alias{codebook.meta.data}
\title{Generic function to build a codebook from an R object}
\usage{
codebook(object, file = NULL, replace = FALSE, render = FALSE)

\method{codebook}{meta.data}(object, file = NULL, replace = FALSE,
  render = FALSE)
}
\arguments{
\item{object}{a data object to construct a codebook for}

\item{file}{a character string indicating where the file should be saved, if
NULL default is `codebook`}

\item{replace}{logical, should an existing codebook file be overwritten, default
is FALSE}

\item{render}{logical, should the codebook be opened after it is written? default
is FALSE}
}
\value{
An HTML file with codebook information
}
\description{
Build an HTML codebook from an R object that includes metadata about the
data stored in the object and its creation. For some object types it will
include auto-generated summary information about the data in the fields.

Build a codebook from a meta.data object
}
\author{
Carl Frederick
}

