% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertMeta.R
\name{convertMeta}
\alias{convertMeta}
\title{Function to convert existing data.frame attributes made by common functions}
\usage{
convertMeta(object, from.package = c("foreign", "haven", "readstata13"))
}
\arguments{
\item{object}{A data.frame-type object}

\item{from.package}{A method specifying the function/package that generated
`object`}
}
\value{
a meta.frame object
}
\description{
\code{convertMeta} takes converts a data.frame (or similar) object produced by
some common functions (i.e. the \code{haven} and \code{foreign} packages) to
a meta.frame object and promotes the attributes to the proper meta.data
places.
}
\details{
Currently, the list of compatible functions/packages includes
\itemize{
  \item{\code{foreign}}
  \item{\code{haven}}
  \item{\code{readstata13}}
}
}
\examples{
require(foreign)
tmp <- read.dta("http://www.ats.ucla.edu/stat/stata/dae/crime.dta")
}
\seealso{
\code{\link[foreign]{read.dta}}, \code{\link[foreign]{read.spss}},
         \code{\link[haven]{read_dta}}, \code{\link[haven]{read_spss}},
         \code{\link[haven]{read_sas}}, \code{\link[readstata13]{read.dta13}}
}

