% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\alias{document.data.frame}
\alias{document.data.table}
\title{Generic function to build a meta.data object}
\usage{
document(data, metadata = NULL, sources = NULL, units = NULL,
  labels = NULL, notes = NULL, revisions = NULL, ...)

\method{document}{data.frame}(data, metadata = NULL, sources = NULL,
  units = NULL, labels = NULL, notes = NULL, revisions = NULL, ...)

\method{document}{data.table}(data, metadata = NULL, sources = NULL,
  units = NULL, labels = NULL, notes = NULL, revisions = NULL, ...)
}
\arguments{
\item{data}{a data object to define metadata for}

\item{metadata}{a meta.data object to use for documentation}

\item{sources}{a list of charcter strings describing data sources}

\item{units}{a list of character strings describing units for the columns}

\item{labels}{a list of character strings providing long variable names and labels}

\item{notes}{a list of character strings with additional notes about variables}

\item{revisions}{a list of character strings with information about revisions to a variable}

\item{...}{optional additional parameters to \code{meta.data.summary}}
}
\value{
A \code{\linkS4class{meta.data}} object
}
\description{
Define and describe meta.data for a data object in R. Based on the object
structure this function will automatically define some attributes of the data.

Setting meta.data from a data.frame object

Setting meta.data from a data.table object
}
\details{
The object has the following items
\itemize{
\item{sources - a list of character strings describing data sources}
\item{units - a list of character strings specifying the units of variables}
\item{labels - a generic character string describing an object}
\item{notes - character strings describing any additional details about specific data elements}
\item{revisions - a list of data revisions}
}
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}

