% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaframe.R
\docType{class}
\name{meta.data-class}
\alias{meta.data}
\alias{meta.data-class}
\title{meta.data data documentation class}
\description{
A meta.data object stores metadata about a given data element in R in a structured
set of lists for easy output and analysis. Most commonly meta.data objects are
attached to a data.frame and stored as a meta.frame.
}
\details{
The object has the following items
\itemize{
\item{sources - a list of character strings describing data sources}
\item{units - a list of character strings specifying the units of variables}
\item{labels - a generic character string describing an object}
\item{notes - character strings describing any additional details about specific data elements}
\item{revisions - a list of data revisions}
\item{var_names - a character vector of column names in last known order}
\item{obs_names - a charadcter vector of row names in last known order}
}
}
\section{Objects from the Class}{
 Objects are created by calls to
\code{\link{meta.data}}.
}
\examples{
showClass("meta.data")
methods(class="meta.data")
}
\seealso{
\code{\link{meta.frame}}
}
\keyword{classes}

