% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaframe.R
\name{meta.frame}
\alias{meta.frame}
\title{meta.frame}
\usage{
meta.frame(data, ...)
}
\arguments{
\item{data}{a data.frame}

\item{...}{additional arguments to \code{\link{document}}}
}
\description{
A data.frame with metadata attached to it
}

