% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaframe.R
\docType{class}
\name{metadata-class}
\alias{metadata}
\alias{metadata-class}
\title{metadata data documentation class}
\description{
A metadata object stores metadata about a given data element in R in a structured
set of lists for easy output and analysis. Most commonly metadata objects are
attached to a data.frame and stored as a metaframe.
}
\details{
The object has the following items
\itemize{
\item{sources - a list of character strings describing data sources}
\item{units - a list of character strings specifying the units of variables}
\item{description - a generic character string describing an object}
\item{annotations - character strings describing any additional details about specific data elements}
\item{revisions - a list of data revisions}
}
}
\section{Objects from the Class}{
 Objects are created by calls to
\code{\link{metadata}}.
}
\examples{
showClass("metadata")
methods(class="metadata")
}
\seealso{
\code{\link{metaframe}}
}
\keyword{classes}

