% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaframe.R
\name{metaframe}
\alias{metaframe}
\title{meta.frame}
\usage{
metaframe(data, ...)
}
\arguments{
\item{data}{a data.frame}

\item{...}{additional arguments to \code{\link{document}}}
}
\description{
A data.frame with metadata attached to it
}

