% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadataMethods.R
\name{print.meta.data}
\alias{print.meta.data}
\title{Print a summary of an \code{\linkS4class{meta.data}} object}
\usage{
\method{print}{meta.data}(x, ...)
}
\arguments{
\item{x}{a \code{\linkS4class{meta.data}} object generated by \code{\link{document}}}

\item{...}{optional additional parameters.}
}
\value{
Values to the screen.
}
\description{
Print a summary of an \code{\linkS4class{meta.data}} object
}
\note{
Not complete
}

