% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaframeMethods.R
\name{print.metaframe}
\alias{print.metaframe}
\title{Print a metaframe object}
\usage{
\method{print}{metaframe}(x, ...)
}
\arguments{
\item{x}{An object of class metaframe}

\item{...}{ignored}
}
\description{
This is a function to print a metaframe
}
\details{
If the number of rows in the data frame exceed 10, then only the 
first ten rows will be printed
}

