% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadataMethods.R
\name{print.summary.meta.data}
\alias{print.summary.meta.data}
\title{Print a meta.data summary}
\usage{
\method{print}{summary.meta.data}(x, ...)
}
\arguments{
\item{x}{An object of class meta.data}

\item{...}{ignored}
}
\description{
This is a function to summarize a meta.data
}

