% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine-doco.R
\name{shine_doc}
\alias{shine_doc}
\title{Launch a shiny app to document your data interactively}
\usage{
shine_doc(data)
}
\arguments{
\item{data}{data you want to document}
}
\value{
A shiny app
}
\description{
\code{shine_doc} launches a shiny app that allows you to interactively
document data elements in your dataframe.
}
