% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skel_reader}
\alias{skel_reader}
\title{Read in an external metadata file formatted by skeleton}
\usage{
skel_reader(file)
}
\arguments{
\item{file}{a file path to read in a csv file}
}
\value{
an object of class meta.data
}
\description{
Read in an external metadata file formatted by skeleton
}
\details{
The csv file should match the format of that produced by the appropriate
skeleton method
}

