% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton}
\alias{skeleton}
\alias{skeleton.data.frame}
\title{Generic function to build a skeleton for codebook input from an R object}
\usage{
skeleton(data, file = NULL, replace = FALSE, fileEncoding = "latin1", ...)

\method{skeleton}{data.frame}(data, file = NULL, replace = FALSE,
  fileEncoding = "latin1", ...)
}
\arguments{
\item{data}{a data object to construct a codebook skeleton for}

\item{file}{a character string indicating where the file should be saved, if 
NULL codebook object is printed to the console}

\item{replace}{logical, should an existing codebook skeleton file be 
overwritten, default is FALSE}

\item{fileEncoding}{file encoding to be passed to \code{\link{write.csv}}, 
default is "latin1"}

\item{...}{additional arguments to pass to \code{\link{write.csv}}}
}
\value{
A CSV file
}
\description{
To make metadata input more efficient, this function builds out a custom 
csv file that can be edited in an outside text editor and read back into R 
and then used to construct a metadata object.
}
\author{
Jared Knowles
}

