% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadataMethods.R
\name{str.meta.data}
\alias{str.meta.data}
\title{Display the structure of a meta.data object}
\usage{
\method{str}{meta.data}(object, ...)
}
\arguments{
\item{object}{An object of class meta.data}

\item{...}{ignored}
}
\description{
This is a function to display the internal structure of a meta.data
}

