% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaframeMethods.R
\name{str.meta.frame}
\alias{str.meta.frame}
\title{Display the structure of a metaframe object}
\usage{
\method{str}{meta.frame}(object, ...)
}
\arguments{
\item{object}{An object of class metaframe}

\item{...}{ignored}
}
\description{
This is a function to display the internal structure of a metaframe
}

