% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaframeMethods.R
\name{str.metaframe}
\alias{str.metaframe}
\title{Display the structure of a metaframe object}
\usage{
\method{str}{metaframe}(object, ...)
}
\arguments{
\item{object}{An object of class metaframe}

\item{...}{ignored}
}
\description{
This is a function to display the internal structure of a metaframe
}
