% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadataMethods.R
\name{summary.meta.data}
\alias{summary.meta.data}
\title{Summarize a meta.data}
\usage{
\method{summary}{meta.data}(object, ...)
}
\arguments{
\item{object}{An object of class meta.data}

\item{...}{ignored}
}
\description{
This is a function to summarize a meta.data
}

