% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaframeMethods.R
\name{summary.meta.frame}
\alias{summary.meta.frame}
\title{Summarize a metaframe}
\usage{
\method{summary}{meta.frame}(object, ...)
}
\arguments{
\item{object}{An object of class metaframe}

\item{...}{ignored}
}
\description{
This is a function to summarize a metaframe
}

