% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaframeMethods.R
\name{summary.metaframe}
\alias{summary.metaframe}
\title{Summarize a metaframe}
\usage{
\method{summary}{metaframe}(object, ...)
}
\arguments{
\item{object}{An object of class metaframe}

\item{...}{ignored}
}
\description{
This is a function to summarize a metaframe
}

