% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_note}
\alias{add_note}
\alias{add_note.meta.data}
\alias{add_note.metaframe}
\title{Add annotation information to data in R}
\usage{
add_note(object, note)

\method{add_note}{meta.data}(object, note)

\method{add_note}{metaframe}(object, note)
}
\arguments{
\item{object}{an object to append with meta.data notes}

\item{note}{a character vector or list for the notes}
}
\description{
This function will add annotations to the dataset. Annotations
should be non-data descriptions of dataset features such as observation dates,
log information, or equipment changes in collecting data.
}

