% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaframe.R
\name{document.data.frame}
\alias{document}
\alias{document.data.frame}
\title{Setting meta.data from a data.frame object}
\usage{
\method{document}{data.frame}(data, sources = NULL, units = NULL,
  labels = NULL, notes = NULL, revisions = NULL)

meta.data(data, ...)
}
\arguments{
\item{data}{a data object to define metadata for}

\item{sources}{a list of charcter strings describing data sources}

\item{...}{optional additional parameters.}
}
\value{
A \code{\linkS4class{meta.data}} object
}
\description{
Setting meta.data from a data.frame object

Define and describe meta.data for a data object in R. Based on the object
structure this function will automatically define some attributes of the data.
}
\details{
The object has the following items
\itemize{
\item{sources - a list of character strings describing data sources}
\item{units - a list of character strings specifying the units of variables}
\item{labels - a generic character string describing an object}
\item{notes - character strings describing any additional details about specific data elements}
\item{revisions - a list of data revisions}
}
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}

