% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{meta.summary}
\alias{meta.summary}
\alias{meta.summary.data.frame}
\title{Generic function to build summary data for a meta.data object}
\usage{
meta.summary(object, n, ...)

\method{meta.summary}{data.frame}(object, n = 3, ...)
}
\arguments{
\item{object}{a data object to define metadata for}

\item{n}{number of top and bottom values to retain, default = 3}

\item{...}{additional arguments to pass through}
}
\value{
A \code{list} with summary features for an object
}
\description{
Summarize a data object in R. Based on the object 
structure this function will automatically define some attributes of the data.
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}
