% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaframe.R
\name{meta.summary.data.frame}
\alias{meta.summary}
\alias{meta.summary.data.frame}
\title{Build summary data from a data.frame object}
\usage{
\method{meta.summary}{data.frame}(object, n = 5)

meta.summary(object, ...)
}
\arguments{
\item{object}{a data object to define metadata for}

\item{n}{the number of unique values to report}
}
\value{
A \code{list} with summary features for an object
}
\description{
Build summary data from a data.frame object

Summarize a data object in R. Based on the object
structure this function will automatically define some attributes of the data.
}
\note{
Yadda yadda yadda
}
\author{
Jared E. Knowles
}

