% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaframe-package.R, R/metaframe.R
\docType{package}
\name{metaframe}
\alias{metaframe}
\alias{metaframe-package}
\alias{metaframe-package}
\alias{metaframe}
\title{metaframe: Provides methods for writing, reading, and analyzing metadata 
about data in data.frame objects in R.}
\usage{
metaframe(data, ...)
}
\arguments{
\item{data}{a data.frame}

\item{...}{additional arguments to \code{\link{document}}}
}
\description{
The metaframe package seeks to make storing data about data easy and seamless 
within the R ecosystem. Attributes of data can be stored as data.frame 
attributes making it easy for analysts to track the provenance of data 
through the data analysis process.

A data.frame with metadata attached to it
}
\details{
See the vignettes for usage examples
}
\section{metaframe functions}{


\itemize{
  \item \code{\link{document}}
  \item \code{\link{codebook}}
}
}

