% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaframeMethods.R
\name{print.metaframe}
\alias{print.metaframe}
\title{Print a metaframe object}
\usage{
\method{print}{metaframe}(x, ...)
}
\arguments{
\item{x}{An object of class metaframe}

\item{...}{ignored}
}
\description{
This is a function to print a metaframe
}

