% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skel_reader}
\alias{skel_reader}
\title{Read in an external metadata file formatted by skeleton}
\usage{
skel_reader(file, fileEncoding = "latin1", ...)
}
\arguments{
\item{file}{a file path to read in a csv file}

\item{fileEncoding}{character, encoding of the csv file to read, default is "latin1"}

\item{...}{additional arguments to pass to \code{\link{read.csv}}}
}
\value{
an object of class meta.data
}
\description{
Read in an external metadata file formatted by skeleton
}
\details{
The csv file should match the format of that produced by the appropriate
skeleton method
}

