% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skel_reader}
\alias{skel_reader}
\title{Read in an external metadata file formatted by skeleton}
\usage{
skel_reader(file)
}
\description{
Read in an external metadata file formatted by skeleton
}
\keyword{internal}

