% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton}
\alias{skeleton}
\alias{skeleton.data.frame}
\title{Generic function to build a skeleton for codebook input from an R object}
\usage{
skeleton(data, file, replace)

\method{skeleton}{data.frame}(data, file = NULL, replace = FALSE)
}
\arguments{
\item{data}{a data object to construct a codebook skeleton for}

\item{file}{a character string indicating where the file should be saved, if
NULL default is `codebook`}

\item{replace}{logical, should an existing codebook skeleton file be
overwritten, default is FALSE}
}
\value{
A CSV file
}
\description{
To make metadata input more efficient, this function builds out a custom
csv file that can be edited in an outside text editor and read back into R
and then used to construct a metadata object.

Build a meta.data skeleton from a data.frame
}
\author{
Jared Knowles
}

