% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton.data.frame}
\alias{skeleton.data.frame}
\title{Build a meta.data skeleton from a data.frame}
\usage{
\method{skeleton}{data.frame}(data, file = NULL, replace = FALSE)
}
\description{
Build a meta.data skeleton from a data.frame
}

